requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_s3

